Create ViewPropTypes
