+++ src/components/ViewAccessibility.js
@@ -1,0 +1,27 @@
+module.exports = {
+  AccessibilityTraits: [
+    'none',
+    'button',
+    'link',
+    'header',
+    'search',
+    'image',
+    'selected',
+    'plays',
+    'key',
+    'text',
+    'summary',
+    'disabled',
+    'frequentUpdates',
+    'startsMedia',
+    'adjustable',
+    'allowsDirectInteraction',
+    'pageTurn',
+  ],
+  AccessibilityComponentTypes: [
+    'none',
+    'button',
+    'radiobutton_checked',
+    'radiobutton_unchecked',
+  ],
+};
